소스 정리 및 오류 처리

+++ src/main/java/com/project/musinsa/service/CodiService.java
@@ -7,7 +7,10 @@ import com.project.musinsa.api.response.BestItemResponse;
 import com.project.musinsa.api.response.model.BestItemModel;
 import com.project.musinsa.api.response.model.LowestModel;
 import com.project.musinsa.core.code.CategoryCode;
+import com.project.musinsa.core.exception.code.ErrorCode;
+import com.project.musinsa.core.exception.exception.CustomException;
 import com.project.musinsa.core.util.CategoryUtil;
+import com.project.musinsa.core.util.JsonUtil;
 import com.project.musinsa.core.util.NumberUtil;
 import com.project.musinsa.entity.CodiEntity;
 import com.project.musinsa.model.convertor.CodiConvertor;
@@ -15,6 +18,7 @@ import com.project.musinsa.model.dto.CodiModel;
 import com.project.musinsa.repository.CodiRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.lang.reflect.Field;
 import java.util.ArrayList;
@@ -49,63 +53,33 @@ public class CodiService {
         List<CodiModel> models = codiConvertor.toDtoList(codiRepository.findAll());
         CodiModel model = models.stream()
                 .min(Comparator.comparing(CodiModel::getTotalPrice))
-                .orElseThrow(NoSuchElementException::new);
-        JsonObject jo = new JsonObject();
-        JsonObject lowestObject = new JsonObject();
-        lowestObject.addProperty("브랜드", model.getBrand());
-        JsonArray array = new JsonArray();
-        try {
-            Object lowestModel = codiConvertor.toLowest(model);
-            for (Field field : lowestModel.getClass().getDeclaredFields()) {
-                field.setAccessible(true);
-                JsonObject categoryObject = new JsonObject();
-                Object value = field.get(lowestModel);
-                categoryObject.addProperty("카테고리", CategoryCode.fromCode(field.getName()).getName());
-                categoryObject.addProperty("가격", NumberUtil.numberComma((Long) value));
-                array.add(categoryObject);
-            }
-        } catch (IllegalAccessException e) {
-            throw new RuntimeException(e);
-        }
-        lowestObject.add("카테고리", array);
-        lowestObject.addProperty("총액", NumberUtil.numberComma(model.getTotalPrice()));
-        jo.add("최저가", lowestObject);
-        return jo.toString();
+                .orElseThrow(() -> new CustomException(ErrorCode.INTERNAL_SERVER_ERROR));
+        Object lowestModel = codiConvertor.toLowest(model);
+        return JsonUtil.lowestJson(model, lowestModel);
     }
 
     public String getCategoryItems(String categoryName) {
         List<CodiModel> models = codiConvertor.toDtoList(codiRepository.findAll());
         Comparator<CodiModel> comparator = CategoryUtil.getCategoryComparator(categoryName);
-        CodiModel minModel = models.stream().min(comparator).orElseThrow(NoSuchElementException::new);
-        CodiModel maxModel = models.stream().max(comparator).orElseThrow(NoSuchElementException::new);
-        JsonObject jo = new JsonObject();
-        jo.addProperty("카테고리", categoryName);
-        JsonArray minArray = new JsonArray();
-        JsonObject minJo = new JsonObject();
-        minJo.addProperty("브랜드", minModel.getBrand());
-        minJo.addProperty("가격", NumberUtil.numberComma(CategoryUtil.getCategory(minModel, categoryName)));
-        minArray.add(minJo);
-        jo.add("최저가", minArray);
-        JsonArray maxArray = new JsonArray();
-        JsonObject maxJo = new JsonObject();
-        maxJo.addProperty("브랜드", maxModel.getBrand());
-        maxJo.addProperty("가격", NumberUtil.numberComma(CategoryUtil.getCategory(maxModel, categoryName)));
-        maxArray.add(maxJo);
-        jo.add("최고가", maxArray);
-        return jo.toString();
+        CodiModel minModel = models.stream().min(comparator).orElseThrow(() -> new CustomException(ErrorCode.INTERNAL_SERVER_ERROR));
+        CodiModel maxModel = models.stream().max(comparator).orElseThrow(() -> new CustomException(ErrorCode.INTERNAL_SERVER_ERROR));
+        return JsonUtil.categoryItemJson(categoryName, minModel, maxModel);
     }
 
-    public void createCodi(CodiRequest codiRequest) {
+    @Transactional
+    public void createAndModifyCodi(CodiRequest codiRequest) {
         CodiEntity entity = codiConvertor.createToEntity(codiRequest);
+        if (entity == null) {
+            throw new CustomException(ErrorCode.REQUEST_DATA_NULL_POINT);
+        }
         codiRepository.save(entity);
     }
 
-    public void modifyCodi(CodiRequest codiRequest) {
-        CodiEntity entity = codiConvertor.createToEntity(codiRequest);
-        codiRepository.save(entity);
-    }
-
+    @Transactional
     public void removeCodi(CodiRequest codiRequest) {
+        if (codiRequest.getBrand() == null) {
+            throw new CustomException(ErrorCode.REQUEST_DATA_NULL_POINT);
+        }
         codiRepository.deleteById(codiRequest.getBrand());
     }
 }
